test(web): add URL-reset branch-ownership regression; split chip test file; fix script indent
